continue work on versioning and endpoints

+++ src/main/java/uk/gov/justice/digital/hmpps/interventionscatalogue/repository/ProviderRepository.java
@@ -8,5 +8,5 @@ import uk.gov.justice.digital.hmpps.interventionscatalogue.model.Provider;
 import java.util.UUID;
 
 @Repository
-public interface ProviderRepository extends RevisionRepository<Provider, UUID, Integer>, JpaRepository<Provider, UUID> {
+public interface ProviderRepository extends RevisionRepository<Provider, UUID, Long>, JpaRepository<Provider, UUID> {
 }
